Worked on Balloons.cs , ScoreBoard.cs and ...
StringExtensions. Renamed methods checkLeft and similar to CheckLeft,
etc.
Added an if statement to check matrix boundaries, so that it doesn`t
write to the console "Index was out of range" every time recursion goes
to its bottom.
renames methods in balloon.cs to be complient with the convention.
Changed properties in ScoreBoard to have private set, added special word
"this" in the constructor to be more readable, rewrote CompareTo method
in the same class