Changed some variables names and edited  the console output so it doesn't stack too much.
Added a clear to the console, so the actions on the console don't stack
and moved the instructions method to it.
Forgot to mention in the 1-st commit that I changed most of the
variables names as well as the class names and  some of the method
names. Added a method to print the fields. Added an exit command to the
switch case and formatted the code.