Renamed some of the classes; Moved methods from Balloons.cs to their appropriative class. Made the OOP connection
Moved the methods that had to do something with the field  from Balloons
to PlayField.cs and their use is from that class. Moved the methods
that had something to do with the score  board from Balloons to
ScoreBoard.cs. Renamed some variables. (Overall made the OOP
connection). Played with the class, namespace and solution names.
NOTE: TODO at Balloons.cs